+++ packages/loaders/tests/multi-tiff.spec.js
@@ -24,29 +24,27 @@ async function loadImage() {
     imageName: 'tiff-folder',
     tiffs: [
       {
-        name: 'Channel 0',
         selection: { c: 0, t: 0, z: 0 },
         tiff: await (await fromFile(CHANNEL_0_FIXTURE)).getImage(0)
       },
       {
-        name: 'Channel 1',
         selection: { c: 1, t: 0, z: 0 },
         tiff: await (await fromFile(CHANNEL_1_FIXTURE)).getImage(0)
       },
       {
-        name: 'Channel 2',
         selection: { c: 2, t: 0, z: 0 },
         tiff: await (await fromFile(CHANNEL_2_FIXTURE)).getImage(0)
       }
-    ]
+    ],
+    channelNames: ['Channel 0', 'Channel 1', 'Channel 2']
   };
 }
 
 test('Creates correct TiffPixelSource for MultiTIFF.', async t => {
   t.plan(5);
   try {
-    const { imageName, tiffs } = await loadImage();
-    const { data } = await load(imageName, tiffs);
+    const { imageName, tiffs, channelNames } = await loadImage();
+    const { data } = await load(imageName, tiffs, channelNames);
     t.equal(data.length, 1, 'image should not be pyramidal.');
     const [base] = data;
     t.deepEqual(
@@ -73,8 +71,8 @@ test('Creates correct TiffPixelSource for MultiTIFF.', async t => {
 test('Get raster data for MultiTIFF.', async t => {
   t.plan(13);
   try {
-    const { imageName, tiffs } = await loadImage();
-    const { data } = await load(imageName, tiffs);
+    const { imageName, tiffs, channelNames } = await loadImage();
+    const { data } = await load(imageName, tiffs, channelNames);
     const [base] = data;
 
     for (let c = 0; c < 3; c += 1) {
@@ -107,8 +105,8 @@ test('Get raster data for MultiTIFF.', async t => {
 test('Correct MultiTIFF metadata.', async t => {
   t.plan(10);
   try {
-    const { imageName, tiffs } = await loadImage();
-    const { metadata } = await load(imageName, tiffs);
+    const { imageName, tiffs, channelNames } = await loadImage();
+    const { metadata } = await load(imageName, tiffs, channelNames);
     const { Name, Pixels } = metadata;
     t.equal(Name, 'tiff-folder', `Name should be 'tiff-folder'.`);
     t.equal(Pixels.SizeC, 3, 'Should have three channels.');